# info-and-advice-datastore-api/open-api-specification.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Info and Advice Datastore API
paths:
  /api/v0/applications:
    get:
      tags:
        - application
      summary: Get applications
      operationId: getApplications
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 25
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponses"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /api/v0/client-details/{id}:
    get:
      tags:
        - client-details
      summary: Get client details by ID
      operationId: getClientDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetails"
        '401' : { description: 'Unauthorized' }
        '403' : { description: 'Forbidden' }
        '404' : { description: 'Not found' }
        '500' : { description: 'Internal server error' }
  /api/v0/applications/{id}:
    get:
      tags:
        - application
      summary: Get application by ID
      operationId: getApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
  /api/v0/applications/{id}:update-means-data:
    put:
      tags:
        - application
      operationId: updateMeansData
      summary: Update means data for an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
        '500':
          description: Internal server error
  /api/v0/applications:start-case:
    post:
      tags:
        - application
      operationId: startCase
      summary: Start an application/case for a client
      requestBody:
        required: true
        description: Information required to start an application/case
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/StartCaseCommand'
      responses:
        '201' : { description: Created }
        '400' : { description: Bad request }
        '500' : { description: Internal server error }
  /api/v0/applications/{id}:update-evidence:
    put:
      tags:
        - application
      summary: Update evidence for an application
      operationId: updateEvidence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        description: Confirmation of the evidence statement for the application.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvidenceCommand'
      responses:
        '204':
          description: Success
        '400':
          description: Bad request
        '404':
          description: Application not found
        '500':
          description: Internal server error

  /api/v0/applications/{id}/declaration:
    put:
      tags:
        - application
      summary: Update declaration for an application
      operationId: updateDeclaration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        description: Confirmation of the declaration statement for the application.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeclarationCommand'
      responses:
        '204':
          description: Success
        '400':
          description: Bad request
        '404':
          description: Application not found
        '500':
          description: Internal server error

components:
  schemas:
    ClientDetails:
      properties:
        individualLegalAidNumber:
          type: string
          format: uuid
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        niNumber:
          type: string
          example: QQ123456B
          minLength: 9
          maxLength: 9
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time

    Address:
      properties:
        addressLine1:
          type: string
        addressLine2:
          type: string
        addressLine3:
          type: string
        addressLine4:
          type: string
        townOrCity:
          type: string
        postCode:
          type: string
        county:
          type: string
        country:
          type: string
          maxLength: 2
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    
    ApplicationResponse:
      type: object
      required:
        - id
        - individualLegalAidNumber
        - providerFirmId
        - providerOfficeId
        - createdAt
        - createdBy
        - modifiedAt
        - modifiedBy
      properties:
        id:
          type: string
          format: uuid
          example: 
        individualLegalAidNumber:
          type: string
          format: uuid
          example:
          description:
        providerFirmId:
          type: string 
          format: uuid
          example:
          description:
        providerOfficeId:
          type: string 
          format: uuid
          example: 
          description:
        meansAssessmentStatus:
          type: string
          format: uuid
        applicationState:
          $ref: "#/components/schemas/ApplicationState"
        declaration:
          $ref: "#/components/schemas/DeclarationResponse"
        evidence:
          type: object
        reasonForReapplication:
          type: string
        meansAssessmentRequired:
          type: boolean
        typeOfNonMeans:
          type: boolean
        ecfFlag:
          type: boolean
        contribution:
          type: string
        applicationType:
          type: string
        determinationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        modifiedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string

    ApplicationResponses:
      type: array
      items:
        $ref: "#/components/schemas/ApplicationResponse"

    DeclarationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientDeclarationStatus:
          $ref: "#/components/schemas/ClientDeclarationStatus"
        declarationConfirmation:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        modifiedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string

    ClientDeclarationStatus:
      type: string
      enum:
        - DRAFT

    ApplicationState:
      type: string
      enum:
        - DRAFT

    StartCaseCommand:
      type: object
      required:
        - client
        - applicationType
      properties:
        client: 
          $ref: '#/components/schemas/CreateClientCommand'
        applicationType:
          type: string
          enum:
            - RCW
            - CLA
            
    CreateClientCommand:
      type: object
      required:
        - fullName
        - dateOfBirth
        - placeOfBirth
      properties:
        fullName:
          type: string
        dateOfBirth:
           type: string
           format: date
        placeOfBirth:
          type: string
        nationalInsuranceNumber:
          type: string
          pattern: '[A-CEGHJ-NOPR-TW-Z]{2}[0-9]{6}[ABCDs]{1}'
          example: AC123456D
        createAddressCommand:
          $ref: '#/components/schemas/CreateAddressCommand'

    CreateAddressCommand:
      type: object
      required:
        - addressLine1
      properties:
        addressLine1:
          type: string
          example: 10 Downing Street
        addressLine2:
          type: string
        addressLine3:
          type: string
        addressLine4:
          type: string
        townOrCity:
          type: string
          example: London
        postCode:
          type: string 
          example: SW1A 2AA
        county:
          type: string
        country:
          type: string
          maxLength: 2
          example: GB
    DeclarationCommand:
      type: object
      required:
        - declarationConfirmation
      properties:
        declarationConfirmation:
          type: boolean
    UpdateEvidenceCommand:
      type: object
      additionalProperties: true
      minProperties: 1